+++ melvm/src/consts.rs
